Refactor : like페이지 컴포넌트 정리 완료
