+++ src/pages/like/components/Like.styles.ts
