--- src/pages/like/components/Like/Member/index.ts
@@ -1,1 +1,0 @@
-export { Member } from "./Member";
