--- src/pages/like/components/Like/Owner/index.ts
@@ -1,1 +1,0 @@
-export { Owner } from "./Owner";
