--- src/pages/like/components/Like/index.ts
@@ -1,2 +1,0 @@
-export { Member } from "./Member";
-export { Owner } from "./Owner";
