+++ src/pages/like/components/Member/LikeMemberCard/LikeMemberCard.types.ts
