+++ src/pages/like/components/Member/LikeMemberContainer/index.ts
@@ -1,0 +1,1 @@
+export { LikeMemberContainer } from "./LikeMemberContainer";
