+++ src/pages/like/components/Member/index.ts
@@ -1,0 +1,2 @@
+export { LikeMemberCard } from "./LikeMemberCard";
+export { LikeMemberContainer } from "./LikeMemberContainer";
