+++ src/pages/like/components/Owner/LikeOwnerCard/LikeOwnerCard.types.ts
