+++ src/pages/like/components/Owner/LikeOwnerCard/index.ts
