+++ src/pages/like/components/Owner/LikeOwnerContainer/LikeOwnerContainer.tsx
@@ -4,9 +4,9 @@ import { useTranslation } from "next-i18next";
 import { useSelectFollows } from "~/states/server/profile";
 import { FlexColumn, PosCenter, Text } from "~/styles/mixins";
 import { isEmpty } from "~/utils";
-import { LikeOwnerCard } from "./LikeOwnerCard";
+import { LikeOwnerCard } from "../LikeOwnerCard";
 
-export const Owner = () => {
+export const LikeOwnerContainer = () => {
   const user = useUser() as User;
   const { data: follows } = useSelectFollows(user.id);
   const { t } = useTranslation("like");
