+++ src/pages/like/components/Owner/LikeOwnerContainer/index.ts
@@ -1,0 +1,1 @@
+export { LikeOwnerContainer } from "./LikeOwnerContainer";
