+++ src/pages/like/components/Owner/index.ts
@@ -1,0 +1,2 @@
+export { LikeOwnerCard } from "./LikeOwnerCard";
+export { LikeOwnerContainer } from "./LikeOwnerContainer";
